refactor: rename to manage_keycloak
